+++ src/client.js
@@ -2,7 +2,6 @@ import EventEmitter2 from 'eventemitter2';
 import * as events from './events';
 import PluginError from './plugin-error';
 import merge from 'merge';
-import get from 'lodash.get';
 import { defaults as _defaults, allowedEncodings } from './defaults';
 
 export default class Client {
@@ -29,14 +28,14 @@ export default class Client {
     return _defaults;
   }
 
-  _callPluginMethod(method, mutableArgIdx, earlyExit, ...args) {
+  async _callPluginMethod(method, mutableArgIdx, earlyExit, ...args) {
     let i = 0;
 
     const shouldContinue = () => (earlyExit ? args[mutableArgIdx] : true);
 
     while (i < this._plugins.length && shouldContinue()) {
       if (this._plugins[i][method]) {
-        args[mutableArgIdx] = this._plugins[i][method](...args);
+        args[mutableArgIdx] = await this._plugins[i][method](...args);
       }
       i += 1;
     }
@@ -50,6 +49,10 @@ export default class Client {
 
   _callOnSuccesses(request, response) {
     return this._callPluginMethod('onSuccess', 1, false, request, response);
+  }
+
+  _callOnCompletes(request, response) {
+    return this._callPluginMethod('onComplete', 1, false, request, response);
   }
 
   _callOnFails(request, response) {
@@ -68,16 +71,20 @@ export default class Client {
     }
   }
 
-  _getRequest(path, options) {
-    let fullPath = path;
+  _getFullPath(path) {
+    let fullPath = path || '';
     if (this.defaults && this.defaults.url) {
-      fullPath = `${this.defaults.url}${this.defaults.sep}${path}`;
+      fullPath = `${this.defaults.url}${this.defaults.sep}${fullPath}`;
     }
+    return fullPath;
+  }
 
+  _getRequest(path, options) {
+    const fullPath = this._getFullPath(path);
     return new Request(fullPath, options);
   }
 
-  fetch(path, options) {
+  async fetch(path, options) {
     let request;
     let onStartError;
 
@@ -89,37 +96,34 @@ export default class Client {
 
     this.eventEmitter.emit(events.REQUEST_START, request);
     try {
-      request = this._callOnStarts(request);
+      request = await this._callOnStarts(request);
     } catch (err) {
       onStartError = err;
     }
 
-    let fetchPromise;
     if (request && !onStartError) {
-      fetchPromise = fetch(request)
-        .then(response => {
-          let mutatedResponse;
-          if (response.status >= 400) {
-            mutatedResponse = this._callOnFails(request, response);
-            this.eventEmitter.emit(events.REQUEST_FAILURE, request, mutatedResponse);
-          } else {
-            mutatedResponse = this._callOnSuccesses(request, response);
-            this.eventEmitter.emit(events.REQUEST_SUCCESS, request, mutatedResponse);
-          }
+      try {
+        const response = await fetch(request);
 
-          return mutatedResponse;
-        })
-        .catch(err => {
-          const mutatedError = this._callOnErrors(request, err);
-          this.eventEmitter.emit(events.REQUEST_ERROR, request, mutatedError);
-          throw mutatedError;
-        });
-    } else {
-      const err = onStartError || new PluginError();
-      fetchPromise = Promise.reject(err);
+        let mutatedResponse = await this._callOnCompletes(request, response);
+
+        if (response.status >= 400) {
+          mutatedResponse = await this._callOnFails(request, response);
+          this.eventEmitter.emit(events.REQUEST_FAILURE, request, mutatedResponse);
+        } else {
+          mutatedResponse = await this._callOnSuccesses(request, response);
+          this.eventEmitter.emit(events.REQUEST_SUCCESS, request, mutatedResponse);
+        }
+
+        return mutatedResponse;
+      } catch (err) {
+        const mutatedError = await this._callOnErrors(request, err);
+        this.eventEmitter.emit(events.REQUEST_ERROR, request, mutatedError);
+        throw mutatedError;
+      }
     }
 
-    return fetchPromise;
+    throw onStartError || new PluginError();
   }
 
   addPlugin(plugin) {
@@ -243,49 +247,54 @@ export default class Client {
     return { body: _body, contentType: _contentType };
   }
 
-  _buildOptionsWithBody(method, body, options) {
+  _buildOptions(method, body, options) {
     let _options = { ...this.defaults[method] };
     _options = merge.recursive(true, _options, options);
 
-    const { body: encodedBody, contentType } = this._encode(body, _options.headers['Content-Type']);
-    if (contentType === false) {
-      delete _options.headers['Content-Type'];
-    } else {
-      _options.headers['Content-Type'] = contentType;
-    }
+    if (body) {
+      const { body: encodedBody, contentType } = this._encode(
+        body,
+        _options.headers['Content-Type']
+      );
 
-    _options.body = encodedBody;
+      if (contentType === false) {
+        delete _options.headers['Content-Type'];
+      } else {
+        _options.headers['Content-Type'] = contentType;
+      }
+
+      _options.body = encodedBody;
+    }
     return _options;
   }
 
   get(path, body, options) {
-    const _options = options || {};
+    const _options = this._buildOptions('get', undefined, options);
     let queryString = '';
-
-    _options.headers = merge(true, _options.headers, get(this.defaults, 'get.headers'));
 
     if (body && Object.keys(body).length) {
       queryString = this.defaults.queryStringifier.bind(this)(body);
     }
 
     _options.method = 'GET';
-    return this.fetch(path + queryString, _options);
+
+    return this.fetch((path || '') + queryString, _options);
   }
 
   post(path, body, options) {
-    const _options = this._buildOptionsWithBody('post', body, options);
+    const _options = this._buildOptions('post', body, options);
     _options.method = 'POST';
     return this.fetch(path, _options);
   }
 
   put(path, body, options) {
-    const _options = this._buildOptionsWithBody('put', body, options);
+    const _options = this._buildOptions('put', body, options);
     _options.method = 'PUT';
     return this.fetch(path, _options);
   }
 
   patch(path, body, options) {
-    const _options = this._buildOptionsWithBody('patch', body, options);
+    const _options = this._buildOptions('patch', body, options);
     _options.method = 'PATCH';
     return this.fetch(path, _options);
   }