+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 CHANGELOG
 =========
+## [v1.1.2](https://github.com/synapsestudios/fetch-client/compare/v1.1.1...v1.1.2) (2019-04-26)
+* fixed bug where multiple refresh requests were made with the same refresh token if simultaneous requests 401
+
+## [v1.1.1](https://github.com/synapsestudios/fetch-client/compare/v1.1.0...v1.1.1) (2019-04-24)
+* fixed bug in oauth plugin that caused an infinite request loop if token refresh failed
+
+## [v1.1.0](https://github.com/synapsestudios/fetch-client/compare/v1.0.4...v1.1.0) (2019-03-20)
+* updated oauth plugin to know about refresh tokens and to automatically fetch new authorization and id tokens
+
 ## [v1.0.4](https://github.com/synapsestudios/fetch-client/compare/v1.0.3...v1.0.4) (2017-04-04)
 
 * Fixed issue where x-www-form-urlencoded and form-data requests were supplying a `Content-Type: 'undefined'` header